requirements.txt
annotated-types==0.7.0
anyio==4.11.0
beautifulsoup4
certifi==2025.8.3
cffi==2.0.0
charset-normalizer==3.4.3
click==8.3.0
colorama==0.4.6
fastapi==0.115.0
h11==0.16.0
httpcore==1.0.9
httpx==0.27.2
idna==3.10
iniconfig==2.1.0
jsonpatch==1.33
jsonpointer==3.0.0
jsonschema>=4.19.0
langchain-core==0.3.78
langgraph==0.2.45
langgraph-checkpoint==2.1.1
langgraph-sdk==0.1.74
langsmith==0.4.32
lxml
numpy==2.1.1
orjson==3.11.3
ormsgpack==1.10.0
packaging==25.0
pluggy==1.6.0
pycparser==2.23
pydantic==2.9.2
pydantic_core==2.23.4
pypdf
pytest==8.3.3
PyYAML==6.0.3
python-docx
requests==2.32.5
requests-toolbelt==1.0.0
sniffio==1.3.1
soundfile==0.12.1
starlette==0.38.6
tenacity==9.1.2
typing_extensions==4.15.0
urllib3==2.5.0
uvicorn==0.30.6
zstandard==0.25.0
